Various changes to the build process and node-static hosting.
I was waiting to commit this because right now the front page looks like a bad pasteboard.co ripoff. I added the background noise to mask the color banding in the gradients, but now it looks even more similar. It's my fault because I haven't thought about the front page enough and I don't "have anything to say" with it. Needless to say I'm going to be changing it quite a bit as I think about it more.