Skip past/before folder implemented.
This implementation is a bit different from that of Sequential 2. In Sequential 2, we always skipped to the next file after/before the current folder. In this version, we also skip into any folders at the same level as the current node. I'm not 100% sure this is better, but it seems so. Should fix the old "skip, skip, whoa, now I'm at the top of the entire hierarchy already" problem.