Remove the entire concept of skip scrolling.
Trying to skip the corners of the image in smartScroll() was inherently at odds with our scaler's goal to make the image as large as possible.

Instead, I'll make the Home and End keys change pages, along with the number pad. Now AlmostFitScaler's minimum size is huge, and smartScroll() will pan across the screen in four big quadrants, which is desirable fairly often.