Add "smarterScroll" algorithm.
This one picks whether to go horizontally first or vertically first based on the dimensions of the current page. I'm not sure this is the best way to do it, but it seems like it should work pretty well for two page spreads and automatic fit mode on the iPhone in portait mode.